now redirects to email verification screen instead of skipping to app
